+++ Framework/Source/ShadingUtils/Shading.slang
@@ -29,7 +29,11 @@
 #ifndef _FALCOR_SHADING_H_
 #define _FALCOR_SHADING_H_
 
-#include "Helpers.h"
+// Make sure we get the macros like `_fn` and `_ref`
+// TODO: just eliminate these since we know this is pure Slang.
+#include "HostDeviceData.h"
+
+__import Helpers;
 
 /*******************************************************************
 Documentation
@@ -46,24 +50,6 @@ Documentation
 /*******************************************************************
 Shading routines
 *******************************************************************/
-
-/**
-This stores the information about the current light source at the shading point.
-This includes a direction from a shading point towards the light,
-radiance emitted from the light souce, which is *received* at the shading point.
-*/
-struct LightAttribs
-{
-    vec3	L;				///< Normalized direction to the light at shading hit
-    float   shadowFactor;   ///< Shadow factor
-    vec3	lightIntensity;	///< Radiance of the emitted light at shading hit
-
-    vec3    P;              ///< Sampled point on the light source
-    vec3    N;              ///< Normal of the sampled point on the light source
-	float   pdf;            ///< Probability density function of sampling the light source
-
-    vec3    points[4];
-};
 
 /**
 This is an output resulting radiance of a full pass over all material layers
@@ -117,9 +103,10 @@ struct ShadingOutput
 };
 
 /* Include all shading routines, including endpoints (camera and light) */
-#include "Cameras.h"
-#include "Lights.h"
-#include "BSDFs.h"
+
+__import Cameras;
+__import Lights;
+__import BSDFs;
 
 /*******************************************************************
 Material shading building blocks
